Arruma o bug da Recyclerview, que era permissão no manifest (cleartext data) insere o botão "cestinha" para inicio da segunda etapa.

+++ app/src/main/java/com/example/trabalhofinaldm/models/Product.kt
@@ -4,7 +4,7 @@ import java.io.Serializable
 class Product (
     var id:Long,
     var name:String,
-    var price: Long
+    var price: Double
 
 ): Serializable{
